Documentation + cleanup for @action.pydantic
