+++ docs/reference/index.rst
@@ -1,3 +1,4 @@
+
 .. _reference:
 
 ========================
@@ -20,5 +21,6 @@ need functionality that is not publicly exposed, please open an issue and we can
     tracking
     visibility
     lifecycle
+    typing
     integrations/index
     telemetry
