+++ docs/reference/tracking.rst
@@ -3,7 +3,7 @@ Tracking
 ========
 
 Reference on the Tracking/Telemetry API.
-Rather, you should use this throug/in conjunction with :py:meth:`burr.core.application.ApplicationBuilder.with_tracker`.
+Rather, you should use this through/in conjunction with :py:meth:`burr.core.application.ApplicationBuilder.with_tracker`.
 
 
 .. autoclass:: burr.tracking.LocalTrackingClient
