+++ examples/typed-state/curls.sh
