+++ examples/youtube-to-social-media-post/server.py
@@ -1,90 +1,45 @@
 import contextlib
-import json
 import logging
+from typing import Optional
 
+import application
 import fastapi
 import uvicorn
-from application import (
-    ApplicationState,
-    SocialMediaPost,
-    build_application,
-    build_streaming_application,
-    build_streaming_application_async,
-)
-from fastapi.responses import StreamingResponse
 
 from burr.core import Application
-from burr.core.action import AsyncStreamingResultContainer, StreamingResultContainer
 
 logger = logging.getLogger(__name__)
 
 # define a global `burr_app` variable
-burr_app: Application[ApplicationState] = None
-# This does streaming, in sync mode
-burr_app_streaming: Application[ApplicationState] = None
+burr_app: Optional[Application] = None
 
-# And this does streaming, in async mode
-burr_app_streaming_async: Application[ApplicationState] = None
 
-DEFAULT_YOUTUBE_URL = "https://www.youtube.com/watch?v=hqutVJyd3TI"
+def get_burr_app() -> Application:
+    """Retrieve the global Burr app."""
+    if burr_app is None:
+        raise RuntimeError("Burr app wasn't instantiated.")
+    return burr_app
 
 
 @contextlib.asynccontextmanager
 async def lifespan(app: fastapi.FastAPI):
     """Instantiate the Burr application on FastAPI startup."""
     # set value for the global `burr_app` variable
-    global burr_app, burr_app_streaming, burr_app_streaming_async
-    burr_app = build_application()
-    burr_app_streaming = build_streaming_application()
-    burr_app_streaming_async = build_streaming_application_async()
+    global burr_app
+    burr_app = application.build_application()
     yield
 
 
 app = fastapi.FastAPI(lifespan=lifespan)
 
 
-@app.get("/social_media_post", response_model=SocialMediaPost)
-def social_media_post(youtube_url: str = DEFAULT_YOUTUBE_URL) -> SocialMediaPost:
+@app.get("/social_media_post")
+def social_media_post(youtube_url: str, burr_app: Application = fastapi.Depends(get_burr_app)):
     """Creates a completion for the chat message"""
     _, _, state = burr_app.run(halt_after=["generate_post"], inputs={"youtube_url": youtube_url})
 
-    return state.data.post
-
-
-@app.get("/social_media_post_streaming_async", response_class=StreamingResponse)
-async def social_media_post_streaming_async(
-    youtube_url: str = DEFAULT_YOUTUBE_URL,
-) -> StreamingResponse:
-    """Creates a completion for the chat message"""
-
-    async def gen():
-        _, streaming_container = await burr_app_streaming_async.astream_result(
-            halt_after=["generate_post"],
-            inputs={"youtube_url": youtube_url},
-        )  # type: ignore
-        streaming_container: AsyncStreamingResultContainer[ApplicationState, SocialMediaPost]
-        async for post in streaming_container:
-            obj = post.model_dump()
-            yield json.dumps(obj)
-
-    return StreamingResponse(gen())
-
-
-@app.get("/social_media_post_streaming", response_class=StreamingResponse)
-def social_media_post_streaming(youtube_url: str = DEFAULT_YOUTUBE_URL) -> StreamingResponse:
-    """Creates a completion for the chat message"""
-
-    def gen():
-        _, streaming_container = burr_app_streaming.stream_result(
-            halt_after=["generate_post"],
-            inputs={"youtube_url": youtube_url},
-        )  # type: ignore
-        streaming_container: StreamingResultContainer[ApplicationState, SocialMediaPost]
-        for post in streaming_container:
-            obj = post.model_dump()
-            yield json.dumps(obj)
-
-    return StreamingResponse(gen())
+    post = state["post"]
+    return {"formatted_post": post.display(), "post": post.model_dump()}
 
 
 if __name__ == "__main__":
