+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "burr"
-version = "0.30.0rc5"
+version = "0.30.0rc6"
 dependencies = [] # yes, there are none
 requires-python = ">=3.9"
 authors = [
